Add Noto Sans Mono as a font import and standardize font weights
Tweaked the font import string to import all weights of Rubik as
well as Noto Sans Mono (which only happened to be the default mono
font on my computer, but I like the way it looks with Rubik).

After doing this, realized that because I was only importing the
base weight of Rubik, all of the bold text had been showing as
"fake bold." *Real* bold with Rubik (weight 700) is too strong for
my taste, so I added a rule to put weight 600 on all <strong>
elements. This can probably be qualified further.