Remove Gemfile.lock from version control
I have flipflopped on this a few times[1], but my latest thinking
is that I have observed that `bundle install` changes the contents
of Gemfile.lock (which was generated using an Arch Linux install)
when run from Ubuntu 22.04. It still installs just fine; it just
qualifies the version of some SASS dependency differently. Since
the CI job that builds the site and publishes it to GitHub pages
runs on an `ubuntu-latest` runner, there's a good chance this
runner doesn't respect the Gemfile.lock, either, in which case what
good is it really doing for me, other than being something I have
to take pains not to `git add`?

[1]: E.g.

- b4d2716 put Gemfile.lock back into version control, organize .gitignore
- 06c981c remove Gemfile.lock