Add assets/main.scss and custom SCSS styles file
The goal of this branch is to move all of my custom CSS into
custom files whose styles are imported *after* the minima base,
rather than editing copies of the minima _sass/minima/_*.scss files
directly. This will make it easy to update minima.

This rather silly commit adds a

	a * {
	    color: red !important;
	}

rule to prove that everything works as intended.

Ref: <https://blog.slowb.ro/dark-theme-for-minima-jekyll/>