Remove minima.scss and put variable definitions in custom styles
This commit demonstrates the process by which you can move styles
into custom styles instead of patching the minima vendored files.
But it also causes the SCSS build to raise a bunch of deprecation
warnings about outdated math syntax--warnings that I fixed in a
past commit patching minima.scss.

This makes me think that now is not the right moment to do this
operation: minima v2.5.1 is quite outdated and I *like* the
patches I have made to it more than I like the idea of being able
to update the base from underneath.

Will revisit this if/when minima v3 comes out.