Yet further tweaks to interplay of bold weight and monospace
This time I am using a SASS @mixin instead of repeating the group
of properties I want on monospace fonts. This fixes the issue
where inline <code> got bolded at weight 600 (as desired) but
bits of code that got a style like .k that is bolded by the syntax
highlighter were bolded to 700 (too much).