Workstation configuration script (#4)
Add a workstation configuration script that pulls in all the Ruby
dependencies needed to build the site locally.

For now, this is Ubuntu-specific, and the purpose is just to ease my
local workflow. Eventually, the goal is to replace the current stack of
third-party GitHub actions with a self-contained Bash script for greater
portability. But there's more work to be done since the version of
`rbenv` that comes with Ubuntu <= 24.04 (the newest version available as
a GitHub action runner) doesn't support Ruby 3.2.2.
https://github.com/maxkapur/maxkapur.github.io/pull/1