Replace literal base URLs with `{{ site.url }}`
Replace a few links that a hardcoded to `maxkapur.com` with Liquid
syntax that produces the same thing. It is still desirable for
these to be absolute (not relative) links because they point to
resources that are hosted on my domain but *not* built with Jekyll,
instead managed from my DNS tools. That's also why we use `site.url`
instead of appending `site.baseurl` (which is supposed to point to
the blog).

I still haven't replaced literal subdomains like `how.maxkapur.com`
and occurences of `maxkapur.com` in my email address.