Enable server-side KaTeX rendering
Use the `kramdown-math-katex` gem to render KaTeX to MathML during
`jekyll build` instead of in the user's browser. This makes the
site more static and responsive at the cost of a slightly longer
build time.

The approach used here is based on the following blog posts:

- https://gendignoux.com/blog/2020/05/23/katex.html
- https://www.xuningyang.com/blog/2021-01-11-katex-with-jekyll/

The Ruby `katex` gem vendors the KaTeX fonts (.woff2) and CSS, but
I had to add some steps in `configure.sh` after `bundle install`
to hardlink those vendored files into the `assets/` directory so
that they are included in the built site.