Abandoning for now
Abandoning this branch for now after seeing how heavy of a
dependency ImageMagick is, and a build error where Inkscape (which
I think is a dependency of ImageMagick?) keeps trying (and failing)
to read files from `/tmp/`. I *think* this might have to do with
installing Inkscape as a snap (for other business), or some other
bad interaction between the conda and host environments.