Update favicon with new brand colors (#29)
Update the top bar of the favicon to reflect the new value of SASS
`$text-color-light`.

Rerun https://simplefavicons.netlify.app/ to generate the various
favicon styles. I experimented with the `jekyll-favicon` plugin
(https://github.com/afaundez/jekyll-favicon), which can automate this
process, but decided it against it to avoid the ImageMagick dependency
and due to a temp files issue that I still haven't debugged.