Use `fonts.css` and `katex.css` from `main.scss` to unify CSS
Now it suffices to include just `main.css` in the <head> element,
so I don't need to remember to put in KaTeX and fonts as well.

This also means we are no longer conditionally including
`katex.css` and instead including it in every page. I am fine with
this because `katex.css` is now locally hosted (instead of coming
from the KaTeX CDN) and probably will be cached anyway.