Remove `feed_meta` and just code the path myself
This is a relic from the old `jekyll-feed` plugin, which we no
longer use because we have a custom `feed.xml` coded in the root
directory. It just so happens that this `feed_meta` invocation was
pointing to the right place, but using `jekyll-feed` just for that
one command is kind of pointless.