Failed attempt
This branch documents a failed attempt to remove the `jekyll-feed`
plugin. We were only using it for the `feed_meta` command, which
is easy to replace with a custom <link> element, as shown in the
previous commit. But even after doing this, Jekyll still does the
"building feed" step because `jekyll-feed` is a dependency of the
minima theme.

The only way to remove `jekyll-feed` will thus be to remove the
dependency on minima entirely.