Ensure bottom of dependency tree is a file task (WIP)
Problem I was running into before: If a file task depends on a
normal task, then rake always reruns the file task (even if the
file is already present). Need to ensure that you always output
files, even dummy ones, to keep track of which tasks have already
been completed.

In this commit, Ctrl+F "conda_env_sentinel" and
"bundle_install_sentinel" to see the pattern that seems to work. I
think we can also have a list of sentinel files to be a little
more robust.